+++ course_grade_attempts.go
@@ -1,11 +1,13 @@
 package blackboard_rest
 
 import (
+  "github.com/jbaxe2/blackboard.rest/api"
   "github.com/jbaxe2/blackboard.rest/course_grade_attempts"
 )
 
 /**
- * The [CourseGradeAttempts] interface...
+ * The [CourseGradeAttempts] interface provides the base interface for interacting
+ * with the REST API's course grade attempts service.
  */
 type CourseGradeAttempts interface {
   GetAttemptFileMetadataList (
@@ -21,4 +23,27 @@ type CourseGradeAttempts interface {
   DownloadAttemptFile (
     courseId string, attemptId string, attemptFileId string,
   ) ([]byte, error)
+}
+
+/**
+ * The [_CourseGradeAttempts] type implements the Course Grade Attempts interface.
+ */
+type _CourseGradeAttempts struct {
+  service api.Service
+
+  CourseGradeAttempts
+}
+
+/**
+ * The [NewCourseGradeAttempts] function creates a new Course Grade Attempts
+ * instance.
+ */
+func NewCourseGradeAttempts (service api.Service) CourseGradeAttempts {
+  if nil == service {
+    return nil
+  }
+
+  return &_CourseGradeAttempts {
+    service: service,
+  }
 }